refactor: extract pincode mod
